Fixes file upload to AWS
It seems that toniecloud at some point switched the way uploads work. The /file POST request now returns an additional field (x-aws-security-token) which must be included in the AWS upload request. Without this field, the AWS API returns a 403 (which is silently ignored).

Add the security header. Also report the content of an API response if no bean is given.